moved the code from BottomDrawer to exist as part in the drawer container
