+++ src/components/BottomDrawer.tsx
@@ -5,14 +5,8 @@ import {
   PanResponder,
   Animated,
   PanResponderGestureState,
-  Text,
-  TouchableOpacity,
-  ScrollView,
 } from "react-native";
-import SearchBars from "./SearchBars";
-import { useCoords } from "../data/CoordsContext";
 import { BottomDrawerStyle } from "../styles/BottomDrawerStyle";
-import { SearchBar } from "react-native-elements";
 
 const { height, width } = Dimensions.get("window");
 const COLLAPSED_HEIGHT = height * 0.1;
@@ -27,58 +21,8 @@ function BottomDrawer({
   children: ReactNode;
   drawerHeight: Animated.Value;
 }) {
-  const { routeData: routeCoordinates, isTransit, setIsTransit } = useCoords();
-  const [htmlInstructions, setHtmlInstructions] = useState<string[]>([]);
   const currentHeightRef = useRef<number>(EXPANDED_HEIGHT);
 
-  useEffect(() => {
-    if (routeCoordinates && routeCoordinates.length > 0) {
-      const instructions = routeCoordinates[0].legs[0].steps.map(
-        (step: any) => {
-          return step.html_instructions
-            .replace(/<[^<>]*>/g, "")
-            .replace(/(?=Destination)/gi, ". ");
-        }
-      );
-
-      let detailedInstructions: string[] = [];
-      const number = routeCoordinates[0].legs[0].steps.length;
-      const instructionsGeneral = routeCoordinates[0].legs[0].steps.map(
-        (step: any) => {
-          return step.html_instructions.replace(/<[^<>]*>/g, "");
-        }
-      );
-      for (let i = 0; i < number; i++) {
-        const detailedHtmlInstructions = routeCoordinates[0]?.legs[0]?.steps[
-          i
-        ]?.steps?.map((step: any) => {
-          return step.html_instructions
-            .replace(/<[^<>]*>/g, "")
-            .replace(/(?=Destination)/gi, ". ");
-        });
-        if (detailedHtmlInstructions == undefined) {
-          detailedInstructions = detailedInstructions.concat(
-            instructionsGeneral[i]
-          );
-        } else {
-          detailedInstructions = detailedInstructions.concat(
-            instructionsGeneral[i]
-          );
-          detailedInstructions = detailedInstructions.concat(
-            detailedHtmlInstructions
-          );
-        }
-      }
-      if (isTransit == true) {
-        setHtmlInstructions(detailedInstructions);
-      } else {
-        setHtmlInstructions(instructions);
-      }
-      console.log(htmlInstructions);
-    } else {
-      setHtmlInstructions([]);
-    }
-  }, [routeCoordinates]);
 
   const drawerState = useRef<number>(1);
 
@@ -186,18 +130,7 @@ function BottomDrawer({
       <View {...panResponder.panHandlers} style={BottomDrawerStyle.dragHandle}>
         <View style={BottomDrawerStyle.dragIndicator} />
       </View>
-      <SearchBars />
-      <View style={BottomDrawerStyle.listContent}>
-        <ScrollView>
-          {htmlInstructions.length > 0 &&
-            htmlInstructions.map((instruction, index) => (
-              <Text key={index} style={BottomDrawerStyle.instructionsList}>
-                {instruction}
-              </Text>
-            ))}
-        </ScrollView>
-      </View>
-      {/* <View style={BottomDrawerStyle.contentContainer}>{children}</View> */}
+      <View style={BottomDrawerStyle.contentContainer}>{children}</View>
     </Animated.View>
   );
 }
